Integrate COOP in PolicyContainer
This adds Cross-Origin Opener Policy to the policy container.
This was previously within [1], which was then split into COOP and COEP
specific CLs.
With this COOP inheritance is handled by the PolicyContainer.
PolicyContainer does not handle the inheritance from the opener during
a new window creation, given that COOP inheritance differs significantly:
 - It inherits from the top frame of the opener, if the opener is
same-origin with its top frame.
Further details in [2].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2490046/.
[2] https://docs.google.com/document/d/1CnG8IWCXencWxmsQMDt9JHUucfpbWXjgknRjquzquV0/edit#

Bug: 1153648
Change-Id: I77bb56bb0d6cd5c8f1523c14487817e970cb0bd2