[composite-bgcolor-animation] Fallback for empty size element
Right now when we have a background color animation on an element
that has empty size, we won't paint the element but we try to start
the element on the compositor. As a result, it won't work because
we don't have an ElementId.

This CL address the problem by make the background color animation
fallback to the main thread.

Bug: 1193757
Change-Id: I1a6b4d91c106ae830550e54f451d0cdf5dea9721