Populate pointerId, pointerType for clicks from touch
Pointer id and pointer type are not available in GestureManager
where we initiate the click from GestureTap events.

Pointer id algorithm is contained in PointerEventFactory. The problem
with surfacing pointer id in GestureManager is made harder because,
when we need to dispatch the click event, the pointer id for the pointer event
sequence that generates the click has already been deleted from
PointerEventFactory.

What I propose is to keep track of the last pointer id based on
IncomingId made up of rawId and pointer type.

Using this approach, when we dispatch click event in GestureManager we
can then use PointerEventManager to access PointerEventFactory to get
the already deleted pointer id for the pointer event sequence that
generated the click event.

In this approach we would need rawId and pointer type in GestureManager.
Pointer type we can get from WebGestureEvent.primary_pointer_type,
however the rawId we do not have access to.
I propose to add primary_pointer_id to WebGestureEvent and populate it
from TouchEvent in GestureProvider to use it as rawId in GestureManager.

Bug: 1150593
TEST: third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_click_is_a_pointerevent.html
Change-Id: I7e921934f99c78298bad96b69aaa163e63c4be29