[composite-bgcolor-animation] Ensure repaint
Currently we rely on HasCurrentBackgroundColorAnimation() !=
other.HasCurrentBackgroundColorAnimation() in the StyleDifference
to trigger a repaint for an element with background color animation.
But this isn't enough, specifically we could have a composited
bgcolor animation attached to an element at Nth frame, but have
a non-compositable animation attached to this element at (N+1)th
frame.

Bug: 1184832
Change-Id: I8b6d0ae366c03e1cd64f75685b2127c664b75939