Consider HasOverflow true as valid scenario for HandleAtomicInline
The table caption element always have 0 available space, since its text
consumes all of it. Hence, adding an empty span with negative margins
may imply the line to overflow. If we add some atomic-inline elements
in such caption, we might need to process them under the scenario of a
NGInlineInfo which has true sa HasOverflow flag.

In the current code, there is an assertion that under kContinue state
and negative remaining_width (implying overflow) it's not valid to have
the line-info's HasOverflow flag to true.

I believe this assertion is contradictory and that this is indeed a
valid scenario, hence, this CL removes such assertion.

Bug: 1197974
Change-Id: I80c48e5dffe5e5ee25fcb4cc15db02d584adc4b9