MSE-in-Workers: Add tests of detach concurrent with appendBuffer
To help find faults like those fixed in crbug 1195441, this change adds
a web test that is like the existing
mediasource-worker-play-terminate-worker case, though instead it
detaches the MediaSource from the element at various times before,
during and after the worker is buffering in small pieces. Further
refinement of the MediaSource readyState for the scenario where a
concurrent "closing has begun" causes exception on attempt to use the
closing MediaSource will be necessary to refine this new test further to
not ignore so many exceptions in the worker thread; TODOs are included
for that in the new test.

BUG=1195441,878133
TEST=mediasource-worker-detach-element.html, confirmed locally in a
release build with DCHECKS enabled to reliably catch the issue in bug
1195441 (with that bug's fix locally reverted for this check).

Change-Id: I3e041d7b06db96d3d976593f20181d073fad1da1