Fix incorrect clipping in 3D cases with SkiaRenderer overlays on Mac.
As in https://crrev.com/420202ad5b0f151e6c82aa9da22c1a41a8d6dcf6 this
flattens a transform before inverting it to map a clip rect, because
what is wanted here is the points with z=0 in the local space that
transform to the x and y of the clip rect boundaries (with some
arbitrary z) in device space.

This fixes the rendering of https://codepen.io/dudleystorey/pen/eyoFf on
Mac with SkiaRenderer enabled.

The test being added is fixed by the patch when viewed manually with:
  ./Content\ Shell --enable-features=UseSkiaRenderer
but does not show the bug when loaded with:
  ./third_party/blink/tools/run_web_tests.py --additional-driver-flag=--enable-features=UseSkiaRenderer
(and I confirmed that the latter does use the Skia renderer).  I don't
know why this is the case.

Bug: 923766
Change-Id: Iae63ee836bb96bb403a120d47ba5a352210035c1